--- src/blink.c
@@ -1,63 +1,0 @@
-/**
- * Copyright (c) 2020 Raspberry Pi (Trading) Ltd.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-#pragma GCC optimize ("O0")
-
-#include "pico/stdlib.h"
-
-// Pico W devices use a GPIO on the WIFI chip for the LED,
-// so when building for Pico W, CYW43_WL_GPIO_LED_PIN will be defined
-#ifdef CYW43_WL_GPIO_LED_PIN
-#include "pico/cyw43_arch.h"
-#endif
-
-#ifndef LED_DELAY_MS
-#define LED_DELAY_MS 1000
-#endif
-
-// Perform initialisation
-int pico_led_init(void) {
-//#if defined(PICO_DEFAULT_LED_PIN)
-    // A device like Pico that uses a GPIO for the LED will define PICO_DEFAULT_LED_PIN
-    // so we can use normal GPIO functionality to turn the led on and off
-    gpio_init(4);
-    gpio_set_dir(4, GPIO_OUT);
-    return PICO_OK;
-//#elif defined(CYW43_WL_GPIO_LED_PIN)
-    // For Pico W devices we need to initialise the driver etc
- //   return cyw43_arch_init();
-//#endif
-}
-
-// Turn the led on or off
-void pico_set_led(bool led_on) {
-//#if defined(PICO_DEFAULT_LED_PIN)
-    // Just set the GPIO on or off
-    gpio_put(4, led_on);
-//#elif defined(CYW43_WL_GPIO_LED_PIN)
-    // Ask the wifi "driver" to set the GPIO on or off
-  //  cyw43_arch_gpio_put(CYW43_WL_GPIO_LED_PIN, led_on);
-//#endif
-}
-
-int main() {
-    int rc = pico_led_init();
-    hard_assert(rc == PICO_OK);
-    while (true) {
-        pico_set_led(true);
-        sleep_ms(LED_DELAY_MS);
-        pico_set_led(false);
-        sleep_ms(LED_DELAY_MS);
-    }
-}
-
-
-//Default register values
-/*
-WAKE_EN all 1
-SLEEP_EN all 1
-ENABLE 0
-*/
