+++ src/blink_interrupt.c
@@ -1,0 +1,38 @@
+#include <stdio.h>
+#include <pico/stdlib.h>
+#include <pico/sync.h>
+
+#ifndef IN_PIN
+#define IN_PIN 5
+#endif
+
+#ifndef OUT_PIN
+#define OUT_PIN 4
+#endif
+
+void irq_callback(uint gpio, uint32_t event_mask)
+{
+    if (gpio != IN_PIN) return;
+    
+    if (event_mask & GPIO_IRQ_EDGE_RISE) {
+        gpio_put(OUT_PIN, true);
+    } else if (event_mask & GPIO_IRQ_EDGE_FALL) {
+        gpio_put(OUT_PIN, false);
+    }
+}
+
+int main() 
+{
+    stdio_init_all();
+
+    gpio_init(IN_PIN);
+    gpio_set_dir(IN_PIN, GPIO_IN);
+
+    gpio_init(OUT_PIN);
+    gpio_set_dir(OUT_PIN, GPIO_OUT);
+    gpio_put(OUT_PIN, 0);
+
+    gpio_set_irq_enabled_with_callback(IN_PIN, GPIO_IRQ_EDGE_RISE | GPIO_IRQ_EDGE_FALL , true, irq_callback);
+    while(1) __wfi();
+    return 0;
+}
